+++ NewWave.Generator/Grooves/GrooveGenerator.cs
@@ -11,7 +11,7 @@ namespace NewWave.Generator.Grooves
 		{
 			var halfSize = false;
 			var maxBeat = songInfo.TimeSignature.BeatCount;
-			if (maxBeat % 2 == 0 && Randomizer.ProbabilityOfTrue(0.75))
+			if (maxBeat % 2 == 0 && Randomizer.ProbabilityOfTrue(0.75f))
 			{
 				halfSize = true;
 				maxBeat = maxBeat / 2;
@@ -26,10 +26,10 @@ namespace NewWave.Generator.Grooves
 			return new Groove(songInfo.TimeSignature, songInfo.Feel, beats);
 		}
 
-		private static List<double> GetBeatPoints(double maxBeat, int feel)
+		private static List<float> GetBeatPoints(float maxBeat, int feel)
 		{
-			var thisBeat = 0.0;
-			var beatPoints = new List<double>();
+			var thisBeat = 0.0f;
+			var beatPoints = new List<float>();
 			while (thisBeat < maxBeat)
 			{
 				var seg = feel == 4
@@ -45,33 +45,33 @@ namespace NewWave.Generator.Grooves
 			return beatPoints;
 		}
 
-		private static List<List<double>> LengthSegments4 => new List<List<double>>
+		private static List<List<float>> LengthSegments4 => new List<List<float>>
 		{
-			new List<double> { 2 },
-			new List<double> { 2, 1, 1 },
-			new List<double> { 1, 2, 1 },
-			new List<double> { 1 },
-			new List<double> { 1, 1 },
-			new List<double> { 0.5, 0.5 },
-			new List<double> { 1.5, 1, 1.5 },
-			new List<double> { 1.5, 1.5 },
-			new List<double> { 0.5, 1, 0.5 },
-			new List<double> { 0.5, 1.5 },
-			new List<double> { 1.5, 1.5, 1 },
-			new List<double> { 0.5, 0.5, 1 },
-			new List<double> { 1, 0.5, 1, 0.5, 1 }
+			new List<float> { 2 },
+			new List<float> { 2, 1, 1 },
+			new List<float> { 1, 2, 1 },
+			new List<float> { 1 },
+			new List<float> { 1, 1 },
+			new List<float> { 0.5f, 0.5f },
+			new List<float> { 1.5f, 1, 1.5f },
+			new List<float> { 1.5f, 1.5f },
+			new List<float> { 0.5f, 1, 0.5f },
+			new List<float> { 0.5f, 1.5f },
+			new List<float> { 1.5f, 1.5f, 1 },
+			new List<float> { 0.5f, 0.5f, 1 },
+			new List<float> { 1, 0.5f, 1, 0.5f, 1 }
 		};
 
-		private static List<List<double>> LengthSegments3 => new List<List<double>>
+		private static List<List<float>> LengthSegments3 => new List<List<float>>
 		{
-			new List<double> { 2 },
-			new List<double> { 2, 1 },
-			new List<double> { 1, 2 },
-			new List<double> { 1 },
-			new List<double> { 1, 1 },
-			new List<double> { 0.33, 0.67 },
-			new List<double> { 0.67, 0.33 },
-			new List<double> { 0.67, 0.67, 0.67 },
+			new List<float> { 2 },
+			new List<float> { 2, 1 },
+			new List<float> { 1, 2 },
+			new List<float> { 1 },
+			new List<float> { 1, 1 },
+			new List<float> { 0.33f, 0.67f },
+			new List<float> { 0.67f, 0.33f },
+			new List<float> { 0.67f, 0.67f, 0.67f },
 		};
 	}
 }
