+++ NewWave.Test/GeneratorTests/FillTests.cs
@@ -15,7 +15,7 @@ namespace NewWave.Test.GeneratorTests
 		[TestMethod]
 		public void GetFill()
 		{
-			const double lengthInBeats = 4.0;
+			const float lengthInBeats = 4f;
 			Console.WriteLine(TabWriter.AsTab(3, FillGenerator.GetFill(0, lengthInBeats, 3).ToList(), (int)lengthInBeats));
 		}
 
@@ -23,13 +23,13 @@ namespace NewWave.Test.GeneratorTests
 		public void GrooveWithFill()
 		{
 			var time = TimeSignature.CommonTime;
-			var feel = Randomizer.ProbabilityOfTrue(0.6) ? 4 : 3;
+			var feel = Randomizer.ProbabilityOfTrue(0.6f) ? 4 : 3;
 			const int measures = 4;
 			var totalLength = measures * time.BeatCount;
 			var drumStyle = new DrumStyle(SectionType.None);
 			drumStyle.Generate(GrooveGenerator.GenerateGroove(new SongInfo(time, feel, 0, 0)));
 			var grooveNotes = Enumerable.Range(0, measures).SelectMany(i => drumStyle.Notes.Select(n => new PercussionNote(n.Start + i * time.BeatCount, n.PercussionId, n.Velocity)));
-			var fillLength = Randomizer.ProbabilityOfTrue(0.5) ? 2 : 4;
+			var fillLength = Randomizer.ProbabilityOfTrue(0.5f) ? 2 : 4;
 			var fill = FillGenerator.GetFill(totalLength - fillLength, fillLength, feel);
 			grooveNotes = grooveNotes.Where(n => n.Start < totalLength - fillLength).Union(fill);
 			Console.WriteLine(TabWriter.AsTab(feel, grooveNotes.ToList(), measures * time.BeatCount));
