+++ NewWave.Test/GeneratorTests/GeneratorTests.cs
@@ -18,12 +18,12 @@ namespace NewWave.Test.GeneratorTests
 		private const OctavePitch LowestPitch = OctavePitch.E2;
 		private static readonly MarkovGeneratorParameters Parameters = new MarkovGeneratorParameters
 		{
-			MinorKeyFunc = () => new List<OctavePitch> { LowestPitch, LowestPitch + 5, LowestPitch + 2 }[Randomizer.GetWeightedIndex(new List<double>
+			MinorKeyFunc = () => new List<OctavePitch> { LowestPitch, LowestPitch + 5, LowestPitch + 2 }[Randomizer.GetWeightedIndex(new List<float>
 			{
-				0.5, 0.3, 0.2
+				0.5f, 0.3f, 0.2f
 			})],
-			TimeSignatureFunc = () => new TimeSignature(Randomizer.ProbabilityOfTrue(0.75) ? 4 : 3, 4),
-			FeelFunc = t => Randomizer.ProbabilityOfTrue(t.BeatCount == 4 ? 0.65 : 0.8) ? 4 : 3,
+			TimeSignatureFunc = () => new TimeSignature(Randomizer.ProbabilityOfTrue(0.75f) ? 4 : 3, 4),
+			FeelFunc = t => Randomizer.ProbabilityOfTrue(t.BeatCount == 4 ? 0.65f : 0.8f) ? 4 : 3,
 			ChordProgressionFilter = MinorOrDiminshedFilter
 		};
 
@@ -73,7 +73,7 @@ namespace NewWave.Test.GeneratorTests
 			{
 				return n =>
 					n.Data.Quality != ChordQuality.Minor && n.Data.Quality != ChordQuality.Diminished
-						? new MarkovChainNode<Chord>(n.Data, n.Probability * 0.25, n.ChildNodes?.Where(c => c.Probability > 0.08).ToList())
+						? new MarkovChainNode<Chord>(n.Data, n.Probability * 0.25f, n.ChildNodes?.Where(c => c.Probability > 0.08).ToList())
 						: n;
 			}
 		}
